Create ClimbingLeaderboard.md
Alice is playing an arcade game and wants to climb to the top of the leaderboard. Can you help her track her ranking as she beats each level? The game uses Dense Ranking, so its leaderboard works like this:

The player with the highest score is ranked number  on the leaderboard.
Players who have equal scores receive the same ranking number, and the next player(s) receive the immediately following ranking number.
For example, four players have the scores , , , and . Those players will have ranks , , , and , respectively.

When Alice starts playing, there are already  people on the leaderboard. The score of each player  is denoted by . Alice plays for  levels, and we denote her total score after passing each level  as . After completing each level, Alice wants to know her current rank.

You are given an array, , of monotonically decreasing leaderboard scores, and another array, , of Alice's cumulative scores for each level of the game. You must print  integers. The  integer should indicate the current rank of alice after passing the  level.

Input Format

The first line contains an integer, , denoting the number of players already on the leaderboard. 
The next line contains  space-separated integers describing the respective values of . 
The next line contains an integer, , denoting the number of levels Alice beats. 
The last line contains  space-separated integers describing the respective values of .

Constraints

 for 
 for 
The existing leaderboard, , is in descending order.
Alice's scores are cumulative, so  is in ascending order.
Subtask

For  of the maximum score:

Output Format

Print  integers. The  integer should indicate the rank of alice after passing the  level.